fixed a little bit user, loan and review

+++ NetworkTechProject/src/main/java/com/project/networktechproject/infrastructure/entity/LoanEntity.java
@@ -10,7 +10,7 @@ public class LoanEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id")
-    private int id;
+    private long id;
     @ManyToOne
     @JoinColumn(name = "book_id")
     private BookEntity bookId;
@@ -26,6 +26,7 @@ public class LoanEntity {
     @Basic
     @Column(name = "return_date")
     private LocalDate returnDate;
+
 
     public BookEntity getBookId() {
         return bookId;
@@ -43,10 +44,10 @@ public class LoanEntity {
         this.userId = userId;
     }
 
-    public int getId() {return id;
+    public long getId() {return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
